refactor(studio): split oversized files to pass 600-line check
